Creation of an application for demo templates and modification of the Pipfile
