+++ Pipfile
@@ -20,6 +20,3 @@ pillow = ">=6.2.0"
 selenium = "*"
 django-model-utils = "*"
 sentry-sdk = "==0.13.2"
-
-[requires]
-python_version = ">=3.5"
