+++ dashboard/templates/registration/login.html
@@ -60,7 +60,7 @@
                                 </form>
                                 <hr>
                                 <div class="text-center">
-                                    <a class="small" href="{% url 'dashboard:password' %}">
+                                    <a class="small" href="#">
                                         {% trans "Forgot Password?" %}
                                     </a>
                                 </div>
