+++ dashboard/urls.py
@@ -14,18 +14,5 @@ urlpatterns = [
     path('soft/add/', views.soft_add, name="soft-add"),
     path('soft/<int:soft_id>/edit/', views.soft_edit, name="soft-edit"),
     path('config/edit/', views.config_edit, name="config-edit"),
-    path('buttons/', views.buttons, name='buttons'),
-    path('cards/', views.cards, name='cards'),
-    path('colors/', views.colors, name='colors'),
-    path('border/', views.border, name='border'),
-    path('animation/', views.animation, name='animation'),
-    path('other/', views.other, name='other'),
-    path('login/', views.login, name='login'),
     path('register/', views.register, name='register'),
-    path('password/', views.forgot_pwd, name='password'),
-    path('blank/', views.blank, name='blank'),
-    path('err_404/', views.error_404, name='error-404'),
-    path('err_500/', views.error_500, name='error-500'),
-    path('charts/', views.charts, name='charts'),
-    path('table/', views.tables, name='tables'),
 ]
