+++ demo/templates/demo/border.html
