+++ demo/templates/demo/charts.html
