+++ demo/templates/demo/colors.html
