+++ demo/templates/demo/forgot-password.html
@@ -36,10 +36,10 @@
                                 </form>
                                 <hr>
                                 <div class="text-center">
-                                    <a class="small" href="{% url 'dashboard:register' %}">Create an Account!</a>
+                                    <a class="small" href="{% url 'demo:register' %}">Create an Account!</a>
                                 </div>
                                 <div class="text-center">
-                                    <a class="small" href="{% url 'dashboard:login' %}">Already have an account?
+                                    <a class="small" href="{% url 'demo:login' %}">Already have an account?
                                         Login!</a>
                                 </div>
                             </div>
