+++ demo/templates/demo/nav_demo.html
@@ -15,8 +15,8 @@
         <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordionSidebar">
             <div class="bg-white py-2 collapse-inner rounded">
                 <h6 class="collapse-header">Custom Components:</h6>
-                <a class="collapse-item" href="{% url 'dashboard:buttons' %}">Buttons</a>
-                <a class="collapse-item" href="{% url 'dashboard:cards' %}">Cards</a>
+                <a class="collapse-item" href="{% url 'demo:buttons' %}">Buttons</a>
+                <a class="collapse-item" href="{% url 'demo:cards' %}">Cards</a>
             </div>
         </div>
     </li>
@@ -32,10 +32,10 @@
              data-parent="#accordionSidebar">
             <div class="bg-white py-2 collapse-inner rounded">
                 <h6 class="collapse-header">Custom Utilities:</h6>
-                <a class="collapse-item" href="{% url 'dashboard:colors' %}">Colors</a>
-                <a class="collapse-item" href="{% url 'dashboard:border' %}">Borders</a>
-                <a class="collapse-item" href="{% url 'dashboard:animation' %}">Animations</a>
-                <a class="collapse-item" href="{% url 'dashboard:other' %}">Other</a>
+                <a class="collapse-item" href="{% url 'demo:colors' %}">Colors</a>
+                <a class="collapse-item" href="{% url 'demo:border' %}">Borders</a>
+                <a class="collapse-item" href="{% url 'demo:animation' %}">Animations</a>
+                <a class="collapse-item" href="{% url 'demo:other' %}">Other</a>
             </div>
         </div>
     </li>
@@ -56,28 +56,28 @@
              data-parent="#accordionSidebar">
             <div class="bg-white py-2 collapse-inner rounded">
                 <h6 class="collapse-header">Login Screens:</h6>
-                <a class="collapse-item" href="{% url 'dashboard:login' %}">Login</a>
-                <a class="collapse-item" href="{% url 'dashboard:register' %}">Register</a>
-                <a class="collapse-item" href="{% url 'dashboard:password' %}">Forgot Password</a>
+                <a class="collapse-item" href="{% url 'demo:login' %}">Login</a>
+                <a class="collapse-item" href="{% url 'demo:register' %}">Register</a>
+                <a class="collapse-item" href="{% url 'demo:password' %}">Forgot Password</a>
                 <div class="collapse-divider"></div>
                 <h6 class="collapse-header">Other Pages:</h6>
-                <a class="collapse-item" href="{% url 'dashboard:error-404' %}">404 Page</a>
-                <a class="collapse-item" href="{% url 'dashboard:error-500' %}">500 Page</a>
-                <a class="collapse-item active" href="{% url 'dashboard:blank' %}">Blank Page</a>
+                <a class="collapse-item" href="{% url 'demo:error-404' %}">404 Page</a>
+                <a class="collapse-item" href="{% url 'demo:error-500' %}">500 Page</a>
+                <a class="collapse-item active" href="{% url 'demo:blank' %}">Blank Page</a>
             </div>
         </div>
     </li>
 
     <!-- Nav Item - Charts -->
     <li class="nav-item">
-        <a class="nav-link" href="{% url 'dashboard:charts' %}">
+        <a class="nav-link" href="{% url 'demo:charts' %}">
             <i class="fas fa-fw fa-chart-area"></i>
             <span>Charts</span></a>
     </li>
 
     <!-- Nav Item - Tables -->
     <li class="nav-item">
-        <a class="nav-link" href="{% url 'dashboard:tables' %}">
+        <a class="nav-link" href="{% url 'demo:tables' %}">
             <i class="fas fa-fw fa-table"></i>
             <span>Tables</span></a>
     </li>
