+++ demo/templates/demo/register.html
@@ -60,10 +60,10 @@
                                 </form>
                                 <hr>
                                 <div class="text-center">
-                                    <a class="small" href="{% url 'dashboard:password' %}">Forgot Password?</a>
+                                    <a class="small" href="{% url 'demo:password' %}">Forgot Password?</a>
                                 </div>
                                 <div class="text-center">
-                                    <a class="small" href="{% url 'dashboard:login' %}">Already have an account?
+                                    <a class="small" href="{% url 'demo:login' %}">Already have an account?
                                         Login!</a>
                                 </div>
                             </div>
