+++ sbadmin/settings/__init__.py
@@ -34,6 +34,7 @@ INSTALLED_APPS = [
     'raspeedi.apps.RaspeediConfig',
     'squalaetp.apps.SqualaetpConfig',
     'reman.apps.RemanConfig',
+    'demo.apps.DemoConfig',
     'api.apps.ApiConfig',
     'django.contrib.admin',
     'django.contrib.auth',
