+++ sbadmin/urls.py
@@ -30,6 +30,7 @@ urlpatterns = [
     path('accounts/', include('django.contrib.auth.urls')),
     path('i18n/', include('django.conf.urls.i18n')),
     path('admin/', admin.site.urls),
+    path('demo/', include('demo.urls')),
 ]
 
 if settings.DEBUG:
